Python Scripts/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
numpy==1.26.3
scipy==1.12.0
openai-whisper==20231117
torch==2.1.2
